refactor(repo): split log.go into queries/stats/settings/cleanup/insert files
